warnings

+++ includes/structure_clone.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   structure_clone.h                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lbougero <lbougero@student.42.fr>          +#+  +:+       +#+        */
+/*   By: hsabouri <hsabouri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/22 15:58:57 by hsabouri          #+#    #+#             */
-/*   Updated: 2019/04/26 13:35:30 by lbougero         ###   ########.fr       */
+/*   Updated: 2019/05/15 11:47:49 by hsabouri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -135,7 +135,7 @@ typedef struct		s_c_point
 typedef struct		s_c_portal
 {
 	size_t		magic;
-	ssize_t		blocking;
+	int			blocking;
 	u_int32_t	from_sector;
 	u_int32_t	to_sector;
 	u_int32_t	from_wall;
